requirements.txt
langgraph == 0.2.72
langgraph-checkpoint-mongodb == 0.1.4
langchain == 0.3.25
langchain-openai == 0.3.28
langchain-core == 0.3.69
langchain-community == 0.3.24
langchain-chroma == 0.2.4
chromadb == 1.0.15
langchain-huggingface == 0.2.0
sentence-transformers == 3.4.1
fastapi == 0.115.11
python-dotenv == 1.0.1
pymongo == 4.12.1
motor == 3.7.1
pydantic ==  2.10.6
uvicorn[standard] == 0.29.0
